refactor: shared 계층 폴더구조 변경
