+++ apps/blog/app/robots.ts
@@ -1,5 +1,5 @@
 import type { MetadataRoute } from "next";
-import { BASE_SITE_URL } from "../src/shared/constants";
+import { BASE_SITE_URL } from "../src/shared/constants/constants";
 
 export default function robots(): MetadataRoute.Robots {
   return {
