+++ apps/blog/src/entities/post/ui/post/PostDetailAuthorAndDate.tsx
@@ -1,8 +1,8 @@
 import { Image } from "@repo/router";
 import { Flex, Paragraph } from "@xionwcfm/xds";
 import { format } from "date-fns";
-import { ASSET_XION_CIRCLE_LOGO_16_16 } from "~/shared/assets";
-import { AUTHOR_NAME } from "~/shared/constants";
+import { AUTHOR_NAME } from "~/shared/constants/constants";
+import { ASSET_XION_CIRCLE_LOGO_16_16 } from "~/shared/constants/images";
 
 type PostDetailAuthorAndDateProps = {
   date: string;
