+++ apps/blog/src/shared/constants/constants.ts
@@ -1,4 +1,3 @@
-// 메타데이터에 대한 상수값입니다.
 export const BASE_SITE_URL = "https://www.xionwcfm.com";
 
 export const BASE_SITE_DESCRIPTION =
@@ -9,8 +8,9 @@ export const BASE_SITE_NAME = "XionWCFM Frontend Tech Blog";
 export const BASE_SITE_TITLE = "XionWCFM | 유길종 - 프론트엔드 개발자 기술 블로그";
 
 export const BASE_TEMPLETE_TITLE = "XionWCFM | 유길종";
+
 export const AUTHOR_NICKNAME = "XionWCFM";
-// author에 대한 상수값입니다.
+
 export const AUTHOR_NAME = "유길종(XionWCFM)";
 
 export const AUTHOR_DESCRIPTION =
@@ -20,37 +20,8 @@ export const AUTHOR_EMAIL = "giljongyudev@gmail.com";
 
 export const AUTHOR_COPYRIGHT = "© 2024 YuGilJong. All rights reserved";
 
-// 기타 서드파티에 대한 상수값입니다.
 export const LORDICON_LICENSE = "Animated icons by Lordicon.com";
 
 export const EXTERNAL_LINKED_IN = "https://www.linkedin.com/in/giljong-yu-289195277/";
 
 export const EXTERNAL_GITHUB = "https://github.com/XionWCFM";
-
-const GRASSHOPPER_SITE_URL = "https://grasshopper.xionwcfm.com";
-
-const XDS_STORYBOOK_SITE_URL = "https://xionwcfm-lib.vercel.app/?path=/docs/xds-button--docs";
-
-const RESUME_URL = `${BASE_SITE_URL}/resume`;
-
-const TECH_BLOG_URL = "https://xionwcfm.tistory.com/";
-
-export const CONSTANTS = {
-  TECH_BLOG_URL,
-  RESUME_URL,
-  BASE_TEMPLETE_TITLE,
-  BASE_SITE_URL,
-  BASE_SITE_DESCRIPTION,
-  BASE_SITE_NAME,
-  BASE_SITE_TITLE,
-  AUTHOR_NICKNAME,
-  AUTHOR_NAME,
-  AUTHOR_DESCRIPTION,
-  AUTHOR_EMAIL,
-  AUTHOR_COPYRIGHT,
-  LORDICON_LICENSE,
-  EXTERNAL_GITHUB,
-  EXTERNAL_LINKED_IN,
-  GRASSHOPPER_SITE_URL,
-  XDS_STORYBOOK_SITE_URL,
-};
