+++ apps/blog/src/shared/constants/fonts/Pretendard-SemiBold.woff2
