+++ apps/blog/src/shared/ui/icon/XionGithubIcon.tsx
@@ -1,7 +1,7 @@
 import { GitHubLogoIcon } from "@radix-ui/react-icons";
 import { Link } from "@repo/router";
 import { cn } from "@xionwcfm/xds";
-import { EXTERNAL_GITHUB } from "~/shared/constants";
+import { EXTERNAL_GITHUB } from "~/shared/constants/constants";
 import type { IconProps } from "./icon.type";
 
 export const XionGithubLogoIcon = (props: IconProps) => {
