+++ apps/blog/src/shared/ui/icon/XionLinkedInIcon.tsx
@@ -1,7 +1,7 @@
 import { LinkedInLogoIcon } from "@radix-ui/react-icons";
 import { Link } from "@repo/router";
 import { cn } from "@xionwcfm/xds";
-import { EXTERNAL_LINKED_IN } from "~/shared/constants";
+import { EXTERNAL_LINKED_IN } from "~/shared/constants/constants";
 import type { IconProps } from "./icon.type";
 
 export const XionLinkedinIcon = (props: IconProps) => {
