+++ apps/blog/src/widgets/header/static-header.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { Image, Link } from "@repo/router";
 import { Flex } from "@xionwcfm/xds";
-import { ASSET_XION_BOX_LOGO_48_16 } from "~/shared/assets";
+import { ASSET_XION_BOX_LOGO_48_16 } from "~/shared/constants/images";
 import { Border } from "~/shared/ui/common/Border";
 
 export const StaticHeader = () => {
